Moved ModalUI into modal_ui.py
