+++ src/rattlesnake/user_interface/abstract_user_interface.py
@@ -7,6 +7,7 @@ from datetime import datetime
 import openpyxl
 
 
+# region: User Interface
 class AbstractUI(ABC):
     """Abstract User Interface class defining the interface with the controller
